Use correct scope for template vars
The vhost template for a puppetmaster using passenger will be evaluated in the
context of apache::vhost, not puppet::master.  This means that it has direct
access to all of the variables in apache::vhost's scope, many of which are
parameterized, and needs fully qualified references back to things that are
only in puppet or puppet::master.  Also, use tabs consistently.